+++ NEWS.md
@@ -1,5 +1,6 @@
 ## xxxx-xx-xx v2.0.0-SNAPSHOT
 * [MODDATAIMP-315](https://issues.folio.org/browse/MODDATAIMP-315) Use Kafka for data-import file processing
+* [MODDATAIMP-352](https://issues.folio.org/browse/MODDATAIMP-352) Implement source reader for EDIFACT files.
 
 ## 2020-11-06 v1.11.1
 * [MODDATAIMP-316](https://issues.folio.org/browse/MODDATAIMP-316) Disable CQL2PgJSON & CQLWrapper extra logging in mod-data-import